+++ DarmuhsTerminalCommands/PatchThings/OtherPatches.cs
@@ -91,6 +91,88 @@ namespace TerminalStuff
         }
     }
 
+    [HarmonyPatch(typeof(FlashlightItem), "Start")]
+    public class Flashlights_Start_Patch
+    {
+        internal static Color? DefaultRegColor { get; private set; }
+        internal static Color? DefaultProColor { get; private set; }
+
+        public static void Postfix(FlashlightItem __instance)
+        {
+            Plugin.Spam($"{__instance.itemProperties.itemName} start!");
+            if (DefaultRegColor.HasValue && __instance.itemProperties.itemName.ToLower() == "flashlight")
+                return;
+
+            if (DefaultProColor.HasValue && __instance.itemProperties.itemName.ToLower() == "pro-flashlight")
+                return;
+
+            if(__instance.itemProperties.itemName.ToLower() == "flashlight")
+                DefaultRegColor = __instance.flashlightBulb.color;
+
+            if (__instance.itemProperties.itemName.ToLower() == "pro-flashlight")
+                DefaultProColor = __instance.flashlightBulb.color;
+
+            Plugin.Spam($"default color has been set for {__instance.itemProperties.itemName}!");
+        }
+    }
+
+    [HarmonyPatch(typeof(FlashlightItem), "SwitchFlashlight")]
+    public class FlashLights_Color_Patch
+    {
+        public static void Postfix(FlashlightItem __instance, bool on)
+        {
+            if (!on)
+                return;
+
+            if(!ConfigSettings.ModNetworking.Value)
+                return;
+
+            if (ColorCommands.RainbowFlash)
+            {
+                NetHandler.Instance.CycleThroughRainbowFlash();
+                return;
+            }
+                
+
+            Color def;
+
+            if (__instance.itemProperties.itemName.ToLower() == "flashlight")
+                def = Flashlights_Start_Patch.DefaultRegColor.Value;
+            else if (__instance.itemProperties.itemName.ToLower() == "pro-flashlight")
+                def = Flashlights_Start_Patch.DefaultProColor.Value;
+            else
+            {
+                def = Color.white;
+                Plugin.Spam($"Unknown flashlight item [ {__instance.itemProperties.itemName} ]");
+            }
+
+            Plugin.Spam($"Color def: {def}\n{__instance.itemProperties.itemName} color: {__instance.flashlightBulb.color}");
+            if (__instance.flashlightBulb.color == def)
+            {
+                if (!ColorCommands.CustomFlashColor.HasValue)
+                {
+                    if (StartOfRound.Instance.localPlayerController.helmetLight.color != def)
+                        NetHandler.Instance.HelmetLightColorServerRpc(def, StartOfRound.Instance.localPlayerController.actualClientId);
+                    return;
+                }
+                    
+
+                Plugin.Spam("Updating from default flashlight color!");
+                NetHandler.SetFlash(ref __instance, ColorCommands.CustomFlashColor.Value);
+                NetHandler.SetHelmetLight(ColorCommands.CustomFlashColor.Value, StartOfRound.Instance.localPlayerController.actualClientId);
+                NetHandler.Instance.FlashColorServerRpc(ColorCommands.CustomFlashColor.Value, StartOfRound.Instance.localPlayerController.actualClientId, StartOfRound.Instance.localPlayerController.playerUsername);
+            }
+            else
+            {
+                if (ColorCommands.CustomFlashColor.HasValue && __instance.bulbLight.color == ColorCommands.CustomFlashColor.Value)
+                    return;
+
+                Plugin.Spam("Updating to new flashlight color!");
+                NetHandler.Instance.HelmetLightColorServerRpc(__instance.bulbLight.color, StartOfRound.Instance.localPlayerController.actualClientId);
+            }
+        }
+    }
+
     public class LoadGrabbablesOnShip
     {
         public static List<GrabbableObject> ItemsOnShip = [];