+++ DarmuhsTerminalCommands/PluginCore/Plugin.cs
@@ -16,7 +16,7 @@ using static OpenLib.ConfigManager.ConfigSetup;
 namespace TerminalStuff
 {
     [BepInPlugin("darmuh.TerminalStuff", "darmuhsTerminalStuff", (PluginInfo.PLUGIN_VERSION))]
-    [BepInDependency("darmuh.OpenLib", "0.2.5")] //hard dependency for my library
+    [BepInDependency("darmuh.OpenLib", "0.2.6")] //hard dependency for my library
 
 
     public class Plugin : BaseUnityPlugin
@@ -26,7 +26,7 @@ namespace TerminalStuff
         {
             public const string PLUGIN_GUID = "darmuh.TerminalStuff";
             public const string PLUGIN_NAME = "darmuhsTerminalStuff";
-            public const string PLUGIN_VERSION = "3.6.7";
+            public const string PLUGIN_VERSION = "3.6.8";
         }
 
         internal static ManualLogSource Log;
@@ -55,11 +55,6 @@ namespace TerminalStuff
         public bool isOnMiniCams = false;
         public bool activeCam = false;
         public bool splitViewCreated = false;
-        public bool escapeConfirmation = false;
-
-        //flashlight stuff
-        public bool fSuccess = false;
-        public bool hSuccess = false;
 
         //AutoComplete
         internal bool removeTab = false;